readme: clarification for AoE timezone
While the language around that timezone tends to be correct in Nix related spaces, it is also not very intuitive IMHO.
The "latest timezone" phrase is technically correct, but hard to parse given the modern usage of "latest" (read: most recent) in tech as the concept does not map 1:1.

With these changes there is a very clear description of "what you see on the clock" vs. "what that means about voting", which should clarify 95% of use-cases, with the last 5% of cases (voters checking within the last 24h) being deferred into a footnote with a shell command that should work virtually everywhere.

Signed-off-by: benaryorg <binary@benary.org>